Added unit test cases for SourcePoint class
